pyproject.toml
[tool.poetry]
name = "bibliotech"
version = "0.1.0"
description = "For communities of book lovers"
authors = ["Kacper Stypik <kacper@stypik.pl>"]
license = "GNU GPL v3"

[tool.poetry.dependencies]
python = "^3.10"
Pillow = "^9.1.0"
Django = "^4.0.4"
djangorestframework = "^3.13.1"
django-environ = "^0.8.1"
djoser = "^2.1.0"
django-autoslug = "^1.9.8"

[tool.poetry.dev-dependencies]
black = "^22.3.0"
psycopg2-binary = "^2.9.3"
pylint-django = "^2.5.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
